requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.6
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.0
pymongo==3.7.2
python-dotenv==0.12.0
pytz==2018.7
six==1.11.0
Werkzeug==1.0.0
